feat(gateway): Fix inferred gateway span design
Move their context and propagator into tracing/api.
Refactor tests
Simplify context element
Fix propagator
