--- components/context/src/test/java/datadog/context/InferredProxyHandlingTest.java
@@ -1,465 +1,0 @@
-package datadog.context;
-
-import static org.junit.jupiter.api.Assertions.assertEquals;
-import static org.junit.jupiter.api.Assertions.assertNotNull;
-import static org.junit.jupiter.api.Assertions.assertNull;
-import static org.junit.jupiter.api.Assertions.assertTrue;
-
-import datadog.context.propagation.CarrierVisitor;
-import datadog.context.propagation.InferredProxyPropagator;
-import java.util.Collections;
-import java.util.HashMap;
-import java.util.Map;
-import java.util.function.BiConsumer;
-import java.util.stream.Stream;
-import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.Nested;
-import org.junit.jupiter.api.Test; // For @Test on nested class methods
-import org.junit.jupiter.params.ParameterizedTest;
-import org.junit.jupiter.params.provider.Arguments;
-import org.junit.jupiter.params.provider.MethodSource;
-
-class InferredProxyHandlingTest {
-
-  // Define header key constants locally for the test
-  static final String PROXY_SYSTEM_KEY = "x-dd-proxy-system";
-  static final String PROXY_REQUEST_TIME_MS_KEY = "x-dd-proxy-request-time-ms";
-  static final String PROXY_PATH_KEY = "x-dd-proxy-path";
-  static final String PROXY_HTTP_METHOD_KEY = "x-dd-proxy-httpmethod";
-  static final String PROXY_DOMAIN_NAME_KEY = "x-dd-proxy-domain-name";
-
-  private InferredProxyPropagator propagator;
-
-  @BeforeEach
-  void setUp() {
-    propagator = new InferredProxyPropagator();
-  }
-
-  // Moved @MethodSource providers to the outer class and made them static
-  static Stream<Arguments> validHeadersProviderForPropagator() {
-    Map<String, String> allStandard = new HashMap<>();
-    allStandard.put(PROXY_SYSTEM_KEY, "aws-apigw"); // The only currently supported system
-    allStandard.put(PROXY_REQUEST_TIME_MS_KEY, "12345");
-    allStandard.put(PROXY_PATH_KEY, "/foo");
-    allStandard.put(PROXY_HTTP_METHOD_KEY, "GET");
-    allStandard.put(PROXY_DOMAIN_NAME_KEY, "api.example.com");
-
-    return Stream.of(
-        Arguments.of(
-            "all standard headers (aws-apigw)",
-            allStandard,
-            "aws-apigw",
-            "12345",
-            "/foo",
-            "GET",
-            "api.example.com",
-            null,
-            null));
-  }
-
-  static Stream<Arguments> invalidOrMissingHeadersProviderForPropagator() { // Renamed
-    Map<String, String> missingSystem = new HashMap<>();
-    missingSystem.put(PROXY_REQUEST_TIME_MS_KEY, "12345");
-    missingSystem.put(PROXY_PATH_KEY, "/foo");
-
-    Map<String, String> missingTime = new HashMap<>();
-    missingTime.put(PROXY_SYSTEM_KEY, "aws-apigw");
-    missingTime.put(PROXY_PATH_KEY, "/foo");
-
-    return Stream.of(
-        Arguments.of("PROXY_SYSTEM_KEY missing", missingSystem),
-        Arguments.of("PROXY_REQUEST_TIME_MS_KEY missing", missingTime));
-  }
-
-  // Simple Map visitor for tests (can remain static or non-static in outer class)
-  static class MapVisitor implements CarrierVisitor<Map<String, String>> {
-    @Override
-    public void forEachKeyValue(Map<String, String> carrier, BiConsumer<String, String> visitor) {
-      if (carrier == null) {
-        return;
-      }
-      carrier.forEach(visitor);
-    }
-  }
-
-  // Custom visitor to test null key path in the extractor - MOVED HERE and made static
-  static class NullKeyTestVisitor implements CarrierVisitor<Map<String, String>> {
-    private final BiConsumer<String, String> actualExtractorAccept;
-
-    NullKeyTestVisitor(BiConsumer<String, String> actualExtractorAccept) {
-      this.actualExtractorAccept = actualExtractorAccept;
-    }
-
-    @Override
-    public void forEachKeyValue(Map<String, String> carrier, BiConsumer<String, String> visitor) {
-      if (actualExtractorAccept != null) {
-        actualExtractorAccept.accept(null, "valueForNullKey");
-      }
-    }
-  }
-
-  @Nested
-  @DisplayName("InferredProxyPropagator Tests")
-  class PropagatorTests { // Kept non-static
-
-    @ParameterizedTest(name = "{0}")
-    @MethodSource(
-        "datadog.context.InferredProxyHandlingTest#validHeadersProviderForPropagator") // Fully
-    // qualified
-    // name
-    @DisplayName("Should extract InferredProxyContext when valid headers are present")
-    void testSuccessfulExtraction(
-        String description,
-        Map<String, String> headers,
-        String expectedSystem,
-        String expectedTimeMs,
-        String expectedPath,
-        String expectedMethod,
-        String expectedDomain,
-        String expectedExtraKey,
-        String expectedExtraValue) {
-
-      Context rootContext = Context.root();
-      // Now accesses the outer class's propagator instance field
-      Context extractedOuterContext = propagator.extract(rootContext, headers, new MapVisitor());
-      InferredProxyContext inferredProxyContext =
-          InferredProxyContext.fromContext(extractedOuterContext);
-
-      assertNotNull(
-          inferredProxyContext, "InferredProxyContext should not be null for: " + description);
-      Map<String, String> actualProxyData = inferredProxyContext.getInferredProxyContext();
-      assertEquals(expectedSystem, actualProxyData.get(PROXY_SYSTEM_KEY));
-      assertEquals(expectedTimeMs, actualProxyData.get(PROXY_REQUEST_TIME_MS_KEY));
-      assertEquals(expectedPath, actualProxyData.get(PROXY_PATH_KEY));
-      assertEquals(expectedMethod, actualProxyData.get(PROXY_HTTP_METHOD_KEY));
-      assertEquals(expectedDomain, actualProxyData.get(PROXY_DOMAIN_NAME_KEY));
-      if (expectedExtraKey != null) {
-        assertEquals(expectedExtraValue, actualProxyData.get(expectedExtraKey));
-      }
-    }
-
-    @ParameterizedTest(name = "{0}")
-    @MethodSource(
-        "datadog.context.InferredProxyHandlingTest#invalidOrMissingHeadersProviderForPropagator") // Fully qualified name
-    @DisplayName("Should create InferredProxyContext even if some critical headers are missing")
-    void testExtractionWithMissingCriticalHeaders(String description, Map<String, String> headers) {
-      Context rootContext = Context.root();
-      Context extractedOuterContext = propagator.extract(rootContext, headers, new MapVisitor());
-      InferredProxyContext inferredProxyContext =
-          InferredProxyContext.fromContext(extractedOuterContext);
-
-      assertNotNull(
-          inferredProxyContext,
-          "InferredProxyContext should still be created if any x-dd-proxy-* header is present for: "
-              + description);
-      Map<String, String> actualProxyData = inferredProxyContext.getInferredProxyContext();
-
-      if (headers.containsKey(PROXY_SYSTEM_KEY)) {
-        assertEquals(headers.get(PROXY_SYSTEM_KEY), actualProxyData.get(PROXY_SYSTEM_KEY));
-      } else {
-        assertNull(actualProxyData.get(PROXY_SYSTEM_KEY));
-      }
-      if (headers.containsKey(PROXY_REQUEST_TIME_MS_KEY)) {
-        assertEquals(
-            headers.get(PROXY_REQUEST_TIME_MS_KEY), actualProxyData.get(PROXY_REQUEST_TIME_MS_KEY));
-      } else {
-        assertNull(actualProxyData.get(PROXY_REQUEST_TIME_MS_KEY));
-      }
-    }
-
-    @Test
-    @DisplayName("Should not extract InferredProxyContext if no relevant headers are present")
-    void testNoRelevantHeaders() {
-      Map<String, String> carrier = new HashMap<>();
-      carrier.put("x-unrelated-header", "value");
-      carrier.put("another-header", "othervalue");
-      Context rootContext = Context.root();
-
-      Context extractedOuterContext = propagator.extract(rootContext, carrier, new MapVisitor());
-      InferredProxyContext inferredProxyContext =
-          InferredProxyContext.fromContext(extractedOuterContext);
-
-      assertNull(
-          inferredProxyContext,
-          "InferredProxyContext should be null if no x-dd-proxy-* headers are found");
-    }
-
-    @Test
-    @DisplayName("Should return original context if carrier is null")
-    void testNullCarrier() {
-      InferredProxyContext initialData =
-          new InferredProxyContext(Collections.singletonMap("test", "value"));
-      Context rootContext = Context.root().with(InferredProxyContext.CONTEXT_KEY, initialData);
-
-      Context extractedOuterContext = propagator.extract(rootContext, null, new MapVisitor());
-
-      assertEquals(rootContext, extractedOuterContext, "Context should be unchanged");
-      assertEquals(
-          "value",
-          InferredProxyContext.fromContext(extractedOuterContext)
-              .getInferredProxyContext()
-              .get("test"));
-    }
-
-    @Test
-    @DisplayName("Should return original context if visitor is null")
-    void testNullVisitor() {
-      Map<String, String> carrier = Collections.singletonMap(PROXY_SYSTEM_KEY, "aws-apigw");
-      InferredProxyContext initialData =
-          new InferredProxyContext(Collections.singletonMap("test", "value"));
-      Context rootContext = Context.root().with(InferredProxyContext.CONTEXT_KEY, initialData);
-
-      Context extractedOuterContext = propagator.extract(rootContext, carrier, null);
-
-      assertEquals(rootContext, extractedOuterContext, "Context should be unchanged");
-      assertEquals(
-          "value",
-          InferredProxyContext.fromContext(extractedOuterContext)
-              .getInferredProxyContext()
-              .get("test"));
-    }
-
-    @Test
-    @DisplayName("Should return original context if context is null")
-    void testNullContext() {
-      Map<String, String> carrier = Collections.singletonMap(PROXY_SYSTEM_KEY, "aws-apigw");
-      Context extractedOuterContext = propagator.extract(null, carrier, new MapVisitor());
-      assertNull(extractedOuterContext, "Context should remain null if passed as null");
-    }
-
-    @Test
-    @DisplayName("Extractor should handle multiple proxy headers")
-    void testMultipleProxyHeaders() {
-      Map<String, String> carrier = new HashMap<>();
-      carrier.put(PROXY_SYSTEM_KEY, "aws-apigw");
-      carrier.put(PROXY_REQUEST_TIME_MS_KEY, "12345");
-      carrier.put("x-dd-proxy-custom", "value1"); // First proxy header
-      carrier.put("x-dd-proxy-another", "value2"); // Second proxy header
-
-      Context rootContext = Context.root();
-      Context extractedOuterContext = propagator.extract(rootContext, carrier, new MapVisitor());
-      InferredProxyContext inferredProxyContext =
-          InferredProxyContext.fromContext(extractedOuterContext);
-
-      assertNotNull(inferredProxyContext);
-      // Check if both headers were stored (covers extractedContext == null being false)
-      assertEquals(
-          "value1", inferredProxyContext.getInferredProxyContext().get("x-dd-proxy-custom"));
-      assertEquals(
-          "value2", inferredProxyContext.getInferredProxyContext().get("x-dd-proxy-another"));
-      assertEquals(
-          "aws-apigw", inferredProxyContext.getInferredProxyContext().get(PROXY_SYSTEM_KEY));
-    }
-
-    @Test
-    @DisplayName("Extractor accept method should handle null/empty keys")
-    void testExtractorAcceptNullEmptyKeys() {
-      Context rootContext = Context.root();
-
-      // Test null key - HashMap doesn't allow null keys. Standard HTTP visitors
-      // also typically don't yield null keys. Testing this branch is difficult
-      // without a custom visitor or modifying the source. Relying on coverage report
-      // or assuming standard carriers won't provide null keys.
-
-      // Test empty key
-      Map<String, String> carrierWithEmptyKey = new HashMap<>();
-      carrierWithEmptyKey.put("", "emptyKeyValue"); // Add empty key
-      carrierWithEmptyKey.put(PROXY_SYSTEM_KEY, "aws-apigw"); // Add a valid key too
-
-      Context contextAfterEmpty =
-          propagator.extract(rootContext, carrierWithEmptyKey, new MapVisitor());
-      InferredProxyContext ipcEmpty = InferredProxyContext.fromContext(contextAfterEmpty);
-
-      // The propagator should ignore the empty key entry entirely.
-      assertNotNull(ipcEmpty, "Context should be created due to valid key");
-      assertNull(ipcEmpty.getInferredProxyContext().get(""), "Empty key should not be stored");
-      assertEquals(
-          "aws-apigw",
-          ipcEmpty.getInferredProxyContext().get(PROXY_SYSTEM_KEY),
-          "Valid key should still be stored");
-      assertEquals(1, ipcEmpty.getInferredProxyContext().size(), "Only valid key should be stored");
-    }
-
-    @Test
-    @DisplayName(
-        "Extractor accept method should handle explicitly passed null key via custom visitor")
-    void testExtractorAcceptExplicitNullKey() {
-      Context rootContext = Context.root();
-      Map<String, String> carrier = new HashMap<>(); // Carrier can be empty for this test
-
-      // We need to get a handle to the internal BiConsumer (the InferredProxyContextExtractor
-      // instance).
-      // The extract method will create one. We can pass a visitor that captures it.
-
-      final BiConsumer<String, String>[] extractorHolder = new BiConsumer[1];
-
-      CarrierVisitor<Map<String, String>> capturingVisitor =
-          (cr, bic) -> {
-            extractorHolder[0] = bic; // Capture the BiConsumer
-            // Optionally, call the original MapVisitor if we still want normal processing after
-            // capture
-            // new MapVisitor().forEachKeyValue(cr, bic);
-          };
-
-      // This first call is primarily to get a reference to the internal extractor
-      propagator.extract(rootContext, carrier, capturingVisitor);
-
-      assertNotNull(extractorHolder[0], "Failed to capture the internal extractor instance");
-
-      // Now use a new custom visitor to specifically test the null key path
-      // on the captured extractor instance (though this isn't how extract is typically used).
-      // A more direct way to test the BiConsumer if it were accessible or if the design allowed it.
-      // For now, we directly call accept on the captured one.
-      extractorHolder[0].accept(null, "valueForNullKey");
-
-      // The goal is JaCoCo coverage. Asserting internal state of the extractor is hard without
-      // reflection.
-      // We can verify that the context remains unchanged or as expected if no valid headers
-      // processed.
-      InferredProxyContext ipc =
-          InferredProxyContext.fromContext(
-              rootContext); // or context returned by a second extract call
-      assertNull(ipc, "Context should not have InferredProxyContext from only a null key call");
-    }
-  }
-
-  @Nested
-  @DisplayName("InferredProxyContext Tests")
-  class ContextUnitTests {
-
-    @Test
-    @DisplayName("Default constructor should create an empty context map")
-    void testDefaultConstructor() {
-      InferredProxyContext ipc = new InferredProxyContext();
-      assertNotNull(ipc.getInferredProxyContext());
-      assertTrue(ipc.getInferredProxyContext().isEmpty());
-    }
-
-    @Test
-    @DisplayName("Constructor with map should initialize context map")
-    void testMapConstructor() {
-      Map<String, String> initialData = new HashMap<>();
-      initialData.put("key1", "value1");
-      initialData.put("key2", "value2");
-
-      InferredProxyContext ipc = new InferredProxyContext(initialData);
-      assertNotNull(ipc.getInferredProxyContext());
-      assertEquals(2, ipc.getInferredProxyContext().size());
-      assertEquals("value1", ipc.getInferredProxyContext().get("key1"));
-      assertEquals("value2", ipc.getInferredProxyContext().get("key2"));
-
-      initialData.put("key3", "value3"); // Modify original map
-      assertNull(ipc.getInferredProxyContext().get("key3"), "Internal map should be a copy");
-    }
-
-    @Test
-    @DisplayName("putInferredProxyInfo should add to the context map")
-    void testPutInfo() {
-      InferredProxyContext ipc = new InferredProxyContext();
-      ipc.putInferredProxyInfo("system", "aws-apigw");
-      ipc.putInferredProxyInfo("time", "12345");
-
-      Map<String, String> contextMap = ipc.getInferredProxyContext();
-      assertEquals(2, contextMap.size());
-      assertEquals("aws-apigw", contextMap.get("system"));
-      assertEquals("12345", contextMap.get("time"));
-
-      ipc.putInferredProxyInfo("system", "azure-func"); // Overwrite
-      assertEquals("azure-func", contextMap.get("system"));
-      assertEquals(2, contextMap.size());
-    }
-
-    @Test
-    @DisplayName("removeInferredProxyInfo should remove from the context map")
-    void testRemoveInfo() {
-      Map<String, String> initialData = new HashMap<>();
-      initialData.put("key1", "value1");
-      initialData.put("key2", "value2");
-      InferredProxyContext ipc = new InferredProxyContext(initialData);
-
-      ipc.removeInferredProxyInfo("key1");
-      Map<String, String> contextMap = ipc.getInferredProxyContext();
-      assertEquals(1, contextMap.size());
-      assertNull(contextMap.get("key1"));
-      assertEquals("value2", contextMap.get("key2"));
-
-      ipc.removeInferredProxyInfo("nonexistent"); // Remove non-existent
-      assertEquals(1, contextMap.size());
-    }
-
-    @Test
-    @DisplayName("storeInto and fromContext should correctly attach and retrieve the context")
-    void testStoreAndFromContext() {
-      InferredProxyContext ipcToStore = new InferredProxyContext();
-      ipcToStore.putInferredProxyInfo("customKey", "customValue");
-
-      Context rootContext = Context.root();
-      Context contextWithValue = ipcToStore.storeInto(rootContext);
-      assertNotNull(contextWithValue);
-
-      InferredProxyContext retrievedIpc = InferredProxyContext.fromContext(contextWithValue);
-      assertNotNull(retrievedIpc);
-      assertEquals("customValue", retrievedIpc.getInferredProxyContext().get("customKey"));
-
-      assertNull(
-          InferredProxyContext.fromContext(rootContext),
-          "Original root context should not be affected");
-
-      Context cleanContext = Context.root();
-      assertNull(
-          InferredProxyContext.fromContext(cleanContext),
-          "fromContext on clean context should be null");
-    }
-
-    @Test
-    @DisplayName("getInferredProxyContext should return an unmodifiable map or a copy")
-    void testGetInferredProxyContextImmutability() {
-      InferredProxyContext ipc = new InferredProxyContext();
-      ipc.putInferredProxyInfo("key1", "value1");
-
-      Map<String, String> retrievedMap = ipc.getInferredProxyContext();
-      assertNotNull(retrievedMap);
-      assertEquals("value1", retrievedMap.get("key1"));
-
-      boolean threwUnsupported = false;
-      try {
-        retrievedMap.put("newKey", "newValue");
-      } catch (UnsupportedOperationException e) {
-        threwUnsupported = true;
-      }
-      // Depending on whether InferredProxyContext.getInferredProxyContext() returns a direct
-      // reference or a copy,
-      // this assertion might change. If it returns a direct mutable reference, threwUnsupported
-      // will be false.
-      // If it returns an unmodifiable view or a copy, attempts to modify might throw or simply not
-      // affect the original.
-      // For now, we check that the original context was not changed.
-      assertEquals(
-          1, ipc.getInferredProxyContext().size(), "Internal map size should remain unchanged");
-      assertEquals(
-          "value1",
-          ipc.getInferredProxyContext().get("key1"),
-          "Internal map content should remain unchanged");
-      // If it MUST be unmodifiable, add: assertTrue(threwUnsupported, "Retrieved map should be
-      // unmodifiable");
-    }
-
-    @Test
-    @DisplayName("Constructor with null map should create an empty context map")
-    void testNullMapConstructor() {
-      InferredProxyContext ipc = new InferredProxyContext(null);
-      assertNotNull(ipc.getInferredProxyContext());
-      assertTrue(ipc.getInferredProxyContext().isEmpty());
-    }
-
-    @Test
-    @DisplayName("Constructor with empty map should create an empty context map")
-    void testEmptyMapConstructor() {
-      Map<String, String> emptyMap = Collections.emptyMap();
-      InferredProxyContext ipc = new InferredProxyContext(emptyMap);
-      assertNotNull(ipc.getInferredProxyContext());
-      assertTrue(ipc.getInferredProxyContext().isEmpty());
-    }
-  }
-}
